Backport fix for bug 3618.
git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/branches/ABI-1-0-0-STABLE@8366 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6

+++ src/af/util/xp/ut_hash.cpp
@@ -618,7 +618,7 @@ const void* UT_StringPtrMap::_next(UT_Cursor& c) const
 const void* UT_StringPtrMap::_prev(UT_Cursor& c) const
 {
 	const hash_slot* map = m_pMapping;
-	size_t x;
+	int x;
 	for (x = c._get_index() - 1; x >= 0; --x)
 	{
 		if (!map[x].empty() && !map[x].deleted())
